Add files via upload
This repository pertains to the exhaustive effort and unwavering dedication that went into completing the machine learning specialization on Coursera. The repository comprises a collection of Jupyter notebooks that have been meticulously documented to showcase the practical applications of several popular algorithms such as linear regression, logistic regression, neural networks, and more. Furthermore, you will find step-by-step explanations of the code, ensuring that you can follow along with my thought process and gain an in-depth understanding of how each algorithm functions, however, replication of the results is not suggested.